requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
gunicorn==23.0.0
langchain==0.3.15
langchain-community==0.3.15
langchain_google_genai==2.0.7
langchain_text_splitters==0.3.5
langchain_chroma==0.1.4
faiss-cpu==1.9.0.post1
google-generativeai==0.8.3
pdfplumber==0.11.0
aiohttp==3.9.5
tqdm==4.66.4
python-dotenv==1.0.1
deep-translator==1.11.4
openpyxl==3.1.2
schedule==1.1.0
